+++ Reqnroll/Configuration/JsonConfig/FormattersConfigurationSourceGenerator.cs
@@ -1,0 +1,15 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace Reqnroll.Configuration.JsonConfig;
+
+[JsonSourceGenerationOptions(WriteIndented = true,
+    PropertyNamingPolicy = JsonKnownNamingPolicy.Unspecified, // We specifiy the names explicitly
+    PropertyNameCaseInsensitive = true, // old custom parser supported ordinal ignore case, so we should do
+    UseStringEnumConverter = true, // use strings instead of numbers for enums
+    ReadCommentHandling = JsonCommentHandling.Skip)] // the user can comment his used configuration value
+[JsonSerializable(typeof(FormattersElement))]
+internal partial class FormattersConfigurationSourceGenerator : JsonSerializerContext
+{
+
+}